requirements.txt
numpy
scipy
opencv-python
pillow
termcolor
pyyaml
tqdm
absl-py
trimesh
torchmetrics
seaborn
tensorboard
scikit-image==0.18.0